backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
anthropic>=0.42.0
claude-agent-sdk
pyyaml>=6.0.1
websockets>=12.0
